requirements.txt
Flask==2.3.2
pytest==7.3.1
gunicorn==20.1.0
python-dotenv==1.0.0
boto3==1.26.157
